+++ dantro/plot/__init__.py
@@ -7,3 +7,4 @@ from .creators import *
 from ..exceptions import SkipPlot
 from .plot_helper import PlotHelper
 from .utils.plot_func import is_plot_func
+from .utils.color_mngr import ColorManager, parse_cmap_and_norm_kwargs